Collapse integration tests into one class.
