+++ tests/Aspire.EndToEnd.Tests/IntegrationServicesTests.cs
@@ -5,14 +5,131 @@ using Xunit;
 
 namespace Aspire.EndToEnd.Tests;
 
-[Collection("IntegrationServices")]
-public class IntegrationServicesTests
+public class IntegrationServicesTests : IClassFixture<IntegrationServicesFixture>
 {
     private readonly IntegrationServicesFixture _integrationServicesFixture;
 
     public IntegrationServicesTests(IntegrationServicesFixture integrationServicesFixture)
     {
         _integrationServicesFixture = integrationServicesFixture;
+    }
+
+    [Fact]
+    public async Task VerifyCosmosWorks()
+    {
+        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(6));
+
+        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", "/cosmos/verify", cts.Token);
+        var responseContent = await response.Content.ReadAsStringAsync();
+
+        Assert.True(response.IsSuccessStatusCode, responseContent);
+    }
+
+    [Fact]
+    public async Task KafkaComponentCanProduceAndConsume()
+    {
+        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(1));
+
+        string topic = $"topic-{Guid.NewGuid()}";
+
+        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", $"/kafka/produce/{topic}", cts.Token);
+        var responseContent = await response.Content.ReadAsStringAsync();
+        Assert.True(response.IsSuccessStatusCode, responseContent);
+
+        response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", $"/kafka/consume/{topic}", cts.Token);
+        responseContent = await response.Content.ReadAsStringAsync();
+        Assert.True(response.IsSuccessStatusCode, responseContent);
+    }
+
+    [Fact]
+    public async Task VerifyMongoWorks()
+    {
+        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(1));
+
+        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", "/mongodb/verify", cts.Token);
+        var responseContent = await response.Content.ReadAsStringAsync();
+
+        Assert.True(response.IsSuccessStatusCode, responseContent);
+    }
+
+    [Fact]
+    public async Task VerifyMySqlWorks()
+    {
+        // MySql health check reports healthy during temporary server phase, c.f. https://github.com/Xabaril/AspNetCore.Diagnostics.HealthChecks/issues/2031
+        // This is mitigated by standard resilience handlers in the IntegrationServicesFixture HttpClient configuration
+
+        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(1));
+
+        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", "/mysql/verify", cts.Token);
+        var responseContent = await response.Content.ReadAsStringAsync();
+
+        Assert.True(response.IsSuccessStatusCode, responseContent);
+    }
+
+    [Fact]
+    public async Task VerifyPomeloEFCoreMySqlWorks()
+    {
+        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(1));
+
+        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", "/pomelo/verify", cts.Token);
+        var responseContent = await response.Content.ReadAsStringAsync();
+
+        Assert.True(response.IsSuccessStatusCode, responseContent);
+    }
+
+    [Fact]
+    public async Task VerifyOracleDatabaseWorks()
+    {
+        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(1));
+
+        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", "/oracledatabase/verify", cts.Token);
+        var responseContent = await response.Content.ReadAsStringAsync();
+
+        Assert.True(response.IsSuccessStatusCode, responseContent);
+    }
+
+    [Fact]
+    public async Task VerifyPostgresWorks()
+    {
+        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(1));
+
+        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", "/postgres/verify", cts.Token);
+        var responseContent = await response.Content.ReadAsStringAsync();
+
+        Assert.True(response.IsSuccessStatusCode, responseContent);
+    }
+
+    [Fact]
+    public async Task VerifyRabbitMQWorks()
+    {
+        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(1));
+
+        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", "/rabbitmq/verify", cts.Token);
+        var responseContent = await response.Content.ReadAsStringAsync();
+
+        Assert.True(response.IsSuccessStatusCode, responseContent);
+    }
+
+    [Fact]
+    public async Task VerifyRedisWorks()
+    {
+        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(1));
+
+        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", "/redis/verify", cts.Token);
+        var responseContent = await response.Content.ReadAsStringAsync();
+
+        Assert.True(response.IsSuccessStatusCode, responseContent);
+    }
+
+    [Fact]
+    public async Task VerifySqlServerWorks()
+    {
+        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(1));
+
+        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", "/sqlserver/verify", cts.Token);
+        var responseContent = await response.Content.ReadAsStringAsync();
+
+        Assert.True(response.IsSuccessStatusCode, responseContent);
     }
 
     [Fact]
