--- tests/Aspire.EndToEnd.Tests/MySql/PomeloEFCoreMySqlFunctionalTests.cs
@@ -1,28 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using Xunit;
-
-namespace Aspire.EndToEnd.Tests.MySql;
-
-[Collection("IntegrationServices")]
-public class PomeloEFCoreMySqlFunctionalTests
-{
-    private readonly IntegrationServicesFixture _integrationServicesFixture;
-
-    public PomeloEFCoreMySqlFunctionalTests(IntegrationServicesFixture integrationServicesFixture)
-    {
-        _integrationServicesFixture = integrationServicesFixture;
-    }
-
-    [Fact]
-    public async Task VerifyPomeloEFCoreMySqlWorks()
-    {
-        using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(1));
-
-        var response = await _integrationServicesFixture.IntegrationServiceA.HttpGetAsync("http", "/pomelo/verify", cts.Token);
-        var responseContent = await response.Content.ReadAsStringAsync();
-
-        Assert.True(response.IsSuccessStatusCode, responseContent);
-    }
-}
